requirements.txt
asgiref==3.5.2
astroid==2.11.7
autopep8==1.6.0
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==4.0.6
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-translate==3.8.1
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.1
openapi-codec==1.3.2
openpyxl==3.0.10
pandas==1.4.3
pathspec==0.9.0
pep8==1.7.1
platformdirs==2.5.2
proto-plus==1.22.0
protobuf==4.21.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyJWT==2.4.0
pylint==2.14.5
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
rsa==4.9
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
